requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
boto3==1.34.0
PyYAML==6.0.1
loguru==0.7.2
